Duplicate list hardening -- JSON unique addition
Experimental change -- I'm not familiar enough with multi-dimensional arrays in Python to be certain of the full effect.

In the precache_equipment call, I added an np.unique/axis=0 to the  main for...in loop.  This should drop duplicates -- Cleaning up the wiki to eliminate duplicates is still required for the correct connections in many cases, but this should avoid breaking the non-duplicate items