Filename sanitization, debug log, enhanced image grab
Filename sanitized to avoid some OS filename issues -- this may be better with a fully featured sanitization library.

A debug log added, default enabled with a .debug [empty file ok] in directory -- will convert to config file with future update

If we do not get a response for the icon grab, remember the failure for the current session but do not save a bad file.  This should cause icon acquisition attempts at each new run, but be fast in the interim.