Update main.py
Typo: "hangars" spelled as "hagars" in one spot, affected some builds.
Repair: on-import repair function added to find this and repair it automatically, throws log repair notice
Some extra checking added in createbutton to look for this problem -- I'm leaving this in at the moment in case of further errors.  It will allow the rest of the app to function and just skip the problem subtypes.